+++ CHANGELOG.md
@@ -8,10 +8,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
 ### Changed
+
 ### Deprecated
+
 ### Removed
+
+- Removed static method `getRoute()`. Use the Router instance with `run()` instead.
+
 ### Fixed
+
 ### Security
 
 ## [2.0.0] - 2025-03-26